requirements.txt
kafka-python==0.9.4
tweepy==3.3.0
configparser==3.3.0r2
